Split os-specific bpf symbolizer to allow some tests on macOS
